requirements.txt
httpx
pytest
pytest-asyncio
pyyaml
json5
tenacity
rich
allpairspy
